+++ packages/shared/src/components/cards/squad/SquadGrid.spec.tsx
@@ -10,31 +10,28 @@ import React from 'react';
 import nock from 'nock';
 import { NextRouter, useRouter } from 'next/router';
 import { mocked } from 'ts-jest/utils';
-import { AuthContextProvider } from '../../../../contexts/AuthContext';
-import loggedUser from '../../../../../__tests__/fixture/loggedUser';
+import { AuthContextProvider } from '../../../contexts/AuthContext';
+import loggedUser from '../../../../__tests__/fixture/loggedUser';
 import {
   generateMembersList,
   generateMembersResult,
   generateTestAdmin,
   generateTestSquad,
-} from '../../../../../__tests__/fixture/squads';
+} from '../../../../__tests__/fixture/squads';
 import { SquadGrid, SourceCardBorderColor } from './SquadGrid';
-import { LazyModalElement } from '../../../modals/LazyModalElement';
+import { LazyModalElement } from '../../modals/LazyModalElement';
 import {
   MockedGraphQLResponse,
   mockGraphQL,
-} from '../../../../../__tests__/helpers/graphql';
+} from '../../../../__tests__/helpers/graphql';
 import {
   SQUAD_JOIN_MUTATION,
   SQUAD_MEMBERS_QUERY,
   SquadEdgesData,
-} from '../../../../graphql/squads';
-import { waitForNock } from '../../../../../__tests__/helpers/utilities';
-import { cloudinary } from '../../../../lib/image';
-import {
-  ActionType,
-  COMPLETE_ACTION_MUTATION,
-} from '../../../../graphql/actions';
+} from '../../../graphql/squads';
+import { waitForNock } from '../../../../__tests__/helpers/utilities';
+import { cloudinary } from '../../../lib/image';
+import { ActionType, COMPLETE_ACTION_MUTATION } from '../../../graphql/actions';
 
 const onClickTest = jest.fn();
 const routerReplace = jest.fn();
