+++ packages/shared/src/components/cards/squad/SquadGrid.tsx
@@ -1,14 +1,14 @@
 import React, { ReactElement } from 'react';
 import classNames from 'classnames';
-import SquadMemberShortList from '../../SquadMemberShortList';
-import { Card } from '../../../cards/Card';
-import { SourceType } from '../../../../graphql/sources';
-import { Image } from '../../../image/Image';
-import { cloudinary } from '../../../../lib/image';
+import SquadMemberShortList from '../../squads/SquadMemberShortList';
+import { Card } from '../Card';
+import { SourceType } from '../../../graphql/sources';
+import { Image } from '../../image/Image';
+import { cloudinary } from '../../../lib/image';
 import { UnFeaturedSquadCardProps } from './common/types';
 import { SquadImage } from './common/SquadImage';
 import { SquadJoinButtonWrapper } from './common/SquadJoinButton';
-import { ButtonVariant } from '../../../buttons/common';
+import { ButtonVariant } from '../../buttons/common';
 
 interface SourceCardProps extends UnFeaturedSquadCardProps {
   borderColor?: SourceCardBorderColor;
