+++ packages/shared/src/components/cards/squad/common/SquadImage.tsx
