+++ packages/shared/src/components/cards/squad/common/SquadJoinButton.tsx
@@ -1,10 +1,10 @@
 import React, { ReactElement } from 'react';
 import { useRouter } from 'next/router';
 import classNames from 'classnames';
-import { SourceType, Squad } from '../../../../../graphql/sources';
-import { SquadJoinButton } from '../../../SquadJoinButton';
-import { Origin } from '../../../../../lib/log';
-import { Button, ButtonVariant } from '../../../../buttons/Button';
+import { SourceType, Squad } from '../../../../graphql/sources';
+import { SquadJoinButton } from '../../../squads/SquadJoinButton';
+import { Origin } from '../../../../lib/log';
+import { Button, ButtonVariant } from '../../../buttons/Button';
 import { SquadCardAction } from './types';
 
 interface SourceJoinButtonProps {
